Entries for Patch 2 (DLC2) - 010055D009F79002
Found these by monitoring the game's memory in Yuzu using cheat engine and watching for a 4-byte integer that changes when changing classes in the certification screen. Eventually I found an address that contains the current loaded palette, and from there I was able to work backwards and infer the IDs for models by comparing the relative offsets in the 'local' IDs (filenames given by extractIndexNum.py) between palettes and models.

I can't guarantee the accuracy of all of these, but I'm confident that the vast majority of them are correct. I've verified most of them in-game by doing replacements using Aldebaran.